Fix use of grep on binary files (#1172405)
os-prober uses 'grep' in an unportable way, in that it assumes that the
regular expression "." matches the NUL byte (all zero bits). POSIX
doesn't guarantee this, and as of grep 2.21 this might not work. If
os-prober assumes GNU grep the fix should be fairly straightforward;
please see the attached patch.

Resolves: rhbz#1172405