Suppress non-blocking dmraid error info (#1198918)
Calls to dmraid did not uniformly redirect standard error. Disks with
unsupported sector sizes are safely ignored and not added to the list of
disks even though an error message is sent to stderr.

Redirected stderr output for invocations of dmraid to /dev/null. This is
safe because actual failures would return a non-zero exit code and
os-prober already guards against those failures.

Resolves: rhbz#1198918